+++ src/components/buttons/index.jsx
@@ -1,4 +1,4 @@
-import LearnMoreBtn from './LearnMoreBtn';
+import LearnMoreBtn from './learnMoreBtn';
 import ReturnHomeBtn from './returnHomeBtn';
 
 export {
